backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.1
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
shap==0.44.0
aiofiles==23.2.1
